updated classes

+++ mativ/lib/model/symbols/value_symbol.dart
@@ -2,11 +2,11 @@ import 'package:mativ/model/operators/operator.dart';
 import 'package:mativ/model/symbols/equation_symbol.dart';
 
 class ValueSymbol extends EquationSymbol {
-  int _value;
+  double _value;
 
-  ValueSymbol(Operator operator, int value) :
+  ValueSymbol(Operator operator, double value) :
     _value = value,
     super(operator);
 
-  int getValue() => _value;
+  double getValue() => _value;
 }
